can validate controller input with Swagger documentation schemas now
